refactor(tagsInput): Changed options initialization
Refactored options initialization code in order to make it simpler and
easier to be extracted into a service in the future so the autocomplete
directive can make use of it.

BREAKING CHANGE: The ngClass option was renamed as customClass so it's
clear that that property is just a value and not the ngClass directive.