fix(tagsInput): Change input width accordingly to its content
Create a new directive to dynamically change the input width so its
content is always visible. For that directive to work, most CSS type
selectors must be changed into class selectors so it's possible to
easily "borrow" the input style and calculate the width of its content.

BREAKING CHANGE: Since CSS selectors must be changed, custom stylesheets
based on the old selectors will conflict with this fix. They'll have to be
updated to use class selectors.

Closes #6.