fix(tagsInput): Remove dependency on interpolation symbols
When the interpolation symbols are changed from the default ones
({{ and }}) the tagsInput directive's template is rendered invalid
since it's using the default interpolation symbols. To fix that, replace
that dependency with a new helper directive, tiBindAttrs, which binds
attributes to properties of the scope and keep those attributes always
up-to-date.

Closes #151