feat(tagsInput): Add type option
Add a new 'type' option so the internal input element can customized.
Only a few values are currently supported: text, email and url. Due to
the use of the ti-bind-attrs directive, Angular will NOT process the
provided value and as a result of that no validation will be performed
whatsoever.

Closes #140.