fix(autocomplete): Fix loadOnEmpty behavior
Fix the behavior of the loadOnEmpty option so the autocomplete list is
loaded truly every time the input field becomes empty. This fix has two
positive effects: 1) the autocomplete list will be also loaded when a tag
is removed and the enableEditingLastTag option is true, and 2) the
autocomplete list won't be hidden when a tag is removed, if it's already
visible.

Closes #205