feat(autoComplete): Add custom template support
Add custom template support for the autoComplete directive so each
suggestion can be freely customized. Introduce a new option called
template that holds either a URL to an HTML file or an id of an inline
script tag. The template, once rendered, is provided with both a special
property containing the matching data and a helper object containing
useful functions (e.g. text highlighting).

Closes #99