feat(tagsInput): Add custom template support
Add custom template support for the tagsInput directive so it matches
the same functionality of the autoComplete directive. Same rules apply
here: tag data is available to the template via the data property,
tag index is available through $index and a tag can be deleted by
calling the $removeTag function.